Refactor: token 엔티티 패키지명 변경
