+++ src/main/java/com/heylocal/traveler/domain/token/AccessToken.java
@@ -1,4 +1,4 @@
-package com.heylocal.traveler.domain.redis;
+package com.heylocal.traveler.domain.token;
 
 import lombok.*;
 import org.springframework.data.annotation.Id;
@@ -7,34 +7,34 @@ import org.springframework.data.redis.core.TimeToLive;
 
 import java.time.LocalDateTime;
 
-@RedisHash("RefreshToken")
+@RedisHash("AccessToken")
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
 @EqualsAndHashCode
 @Builder
-public class RefreshToken {
+public class AccessToken {
   @Id
   private Long userId;
   private String tokenValue;
   @TimeToLive
   private Long timeoutSec;
-  private Long accessTokenId;
+  private Long refreshTokenId;
   @Builder.Default
   private LocalDateTime createdDate = LocalDateTime.now();
 
-  public RefreshToken(Long userId, String tokenValue, Long timeoutSec, Long accessTokenId) {
+  public AccessToken(Long userId, String tokenValue, Long timeoutSec, Long refreshTokenId) {
     this.userId = userId;
     this.tokenValue = tokenValue;
     this.timeoutSec = timeoutSec;
-    this.accessTokenId = accessTokenId;
+    this.refreshTokenId = refreshTokenId;
   }
 
-  public void associateAccessToken(AccessToken accessToken) {
-    this.accessTokenId = accessToken.getUserId();
-    if (accessToken.getRefreshTokenId() != this.getUserId()) {
-      accessToken.associateRefreshToken(this);
+  public void associateRefreshToken(RefreshToken refreshToken) {
+    this.refreshTokenId = refreshToken.getUserId();
+    if (refreshToken.getAccessTokenId() != this.getUserId()) {
+      refreshToken.associateAccessToken(this);
     }
   }
