+++ src/main/java/com/heylocal/traveler/domain/token/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.heylocal.traveler.domain.redis;
+package com.heylocal.traveler.domain.token;
 
 import lombok.*;
 import org.springframework.data.annotation.Id;
@@ -7,34 +7,34 @@ import org.springframework.data.redis.core.TimeToLive;
 
 import java.time.LocalDateTime;
 
-@RedisHash("AccessToken")
+@RedisHash("RefreshToken")
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
 @EqualsAndHashCode
 @Builder
-public class AccessToken {
+public class RefreshToken {
   @Id
   private Long userId;
   private String tokenValue;
   @TimeToLive
   private Long timeoutSec;
-  private Long refreshTokenId;
+  private Long accessTokenId;
   @Builder.Default
   private LocalDateTime createdDate = LocalDateTime.now();
 
-  public AccessToken(Long userId, String tokenValue, Long timeoutSec, Long refreshTokenId) {
+  public RefreshToken(Long userId, String tokenValue, Long timeoutSec, Long accessTokenId) {
     this.userId = userId;
     this.tokenValue = tokenValue;
     this.timeoutSec = timeoutSec;
-    this.refreshTokenId = refreshTokenId;
+    this.accessTokenId = accessTokenId;
   }
 
-  public void associateRefreshToken(RefreshToken refreshToken) {
-    this.refreshTokenId = refreshToken.getUserId();
-    if (refreshToken.getAccessTokenId() != this.getUserId()) {
-      refreshToken.associateAccessToken(this);
+  public void associateAccessToken(AccessToken accessToken) {
+    this.accessTokenId = accessToken.getUserId();
+    if (accessToken.getRefreshTokenId() != this.getUserId()) {
+      accessToken.associateRefreshToken(this);
     }
   }
